+++ pyPalworldAPI/mainapi.py
@@ -7,14 +7,14 @@ import sqlalchemy as sa
 import utils.auth as AuthUtils
 import utils.descriptions as D
 from dotenv import load_dotenv
-from fastapi import FastAPI, status, Request
+from fastapi import FastAPI, Request, status
 from fastapi.responses import JSONResponse
 from fastapi.staticfiles import StaticFiles
 from fastapi_pagination import add_pagination
 from routers import admin, autocomplete, items, misc, oauth2, pals, user
 from sqlmodel.ext.asyncio.session import AsyncSession
-from utils.database import auth_engine, auth_table, engine
 from utils.customexception import APIException
+from utils.database import auth_engine, auth_table, engine
 
 load_dotenv(dotenv_path=".env")
 
@@ -74,13 +74,13 @@ app = FastAPI(
 
 add_pagination(app)
 
+
 @app.exception_handler(APIException)
 async def api_exception_handler(request: Request, exc: APIException):
     return JSONResponse(
-        status_code=exc.status_code,
-        content=exc.content,
-        headers=exc.headers
+        status_code=exc.status_code, content=exc.content, headers=exc.headers
     )
+
 
 if os.getenv("COMPOSE_PROFILES") == "USE_OAUTH2":
     app.include_router(user.router)
@@ -92,7 +92,8 @@ app.include_router(items.router)
 app.include_router(misc.router)
 app.include_router(pals.router)
 
-app.mount("/public", StaticFiles(directory="public"), name="public")
+SCRIPT_DIR = os.path.dirname(os.path.abspath(__file__))
+app.mount("/public", StaticFiles(directory=os.path.join(SCRIPT_DIR, "public")), name="public")
 
 
 @app.get(